Move files to folder and add basic tests for additional post types
