+++ tests/e2e/tests/additional-post-types.spec.ts
@@ -1,0 +1,32 @@
+/**
+ * Import our custom test fixtures.
+ */
+import { test, expect } from '../aql-fixtures';
+
+/**
+ * Internal dependencies.
+ */
+import { insertAQL } from '../utils';
+
+test.describe( 'Additional Post Types', () => {
+	test.beforeEach( async ( { page, editor, playground, admin } ) => {
+		await playground.init( { page, editor } );
+		await admin.visitAdminPage( 'post-new.php' );
+
+		await editor.setPreferences( 'core/edit-post', {
+			welcomeGuide: false,
+			fullscreenMode: false,
+		} );
+		await insertAQL( { editor } );
+	} );
+
+	test.afterEach( async ( { playground } ) => {
+		await playground.cleanUp();
+	} );
+
+	test( 'Control is visible', async ( { page, editor } ) => {
+		expect(
+			page.getByRole( 'combobox', { name: 'Additional Post Types' } )
+		).toBeVisible();
+	} );
+} );
